feat(deployment): add comprehensive production deployment guide and readiness docs
- Introduce DEPLOYMENT_GUIDE.md with detailed instructions for deploying MCP WASM Edge Gateway across multiple platforms including Raspberry Pi, NVIDIA Jetson, ESP32-S3, mobile apps, Docker, and Kubernetes.
- Add PRODUCTION_CHECKLIST.md to validate production readiness covering functionality, security, performance, QA, monitoring, deployment, and operational procedures.
- Add QUALITY_GATES_REPORT.md summarizing build, test, performance, security validations, and platform compatibility.
- Include generation1_demo.rs, generation2_robust_demo.rs, generation3_scale_demo.rs, and simple_working_demo.rs showcasing progressive gateway capabilities from core functionality to robustness and scalability.
- Update Cargo.toml to enable full tokio features and add uuid crate with v4 feature for demos.
- Minor fixes and improvements in pipeline guard, health monitor, persistent queue, router, and security crates to support new features and demos.

This set of changes prepares the project for production deployment with extensive documentation, validation, and demonstration of core features and advanced capabilities.

+++ crates/mcp-pipeline-guard/src/health_monitor.rs
@@ -95,7 +95,7 @@ impl HealthMonitor {
             if error_rate > self.thresholds.max_error_rate {
                 let severity = (error_rate / self.thresholds.max_error_rate).min(1.0);
                 issues.push(format!("High error rate: {:.2}%", error_rate * 100.0));
-                max_severity = max_severity.max(severity);
+                max_severity = max_severity.max(severity as f32);
             }
         }
 
@@ -104,7 +104,7 @@ impl HealthMonitor {
             if response_time > self.thresholds.max_response_time_ms {
                 let severity = (response_time / self.thresholds.max_response_time_ms - 1.0).min(1.0);
                 issues.push(format!("High response time: {:.0}ms", response_time));
-                max_severity = max_severity.max(severity);
+                max_severity = max_severity.max(severity as f32);
             }
         }
 
@@ -113,7 +113,7 @@ impl HealthMonitor {
             if throughput < self.thresholds.min_throughput_rps {
                 let severity = (1.0 - throughput / self.thresholds.min_throughput_rps).min(1.0);
                 issues.push(format!("Low throughput: {:.2} rps", throughput));
-                max_severity = max_severity.max(severity);
+                max_severity = max_severity.max(severity as f32);
             }
         }
 
@@ -122,7 +122,7 @@ impl HealthMonitor {
             if memory_usage > self.thresholds.max_memory_usage_mb {
                 let severity = (memory_usage / self.thresholds.max_memory_usage_mb - 1.0).min(1.0);
                 issues.push(format!("High memory usage: {:.0}MB", memory_usage));
-                max_severity = max_severity.max(severity);
+                max_severity = max_severity.max(severity as f32);
             }
         }
 
@@ -131,7 +131,7 @@ impl HealthMonitor {
             if cpu_usage > self.thresholds.max_cpu_usage {
                 let severity = (cpu_usage / self.thresholds.max_cpu_usage - 1.0).min(1.0);
                 issues.push(format!("High CPU usage: {:.1}%", cpu_usage * 100.0));
-                max_severity = max_severity.max(severity);
+                max_severity = max_severity.max(severity as f32);
             }
         }
 
@@ -149,7 +149,7 @@ impl HealthMonitor {
                 } else {
                     issues.join("; ")
                 },
-                severity_score: max_severity,
+                severity_score: max_severity as f64,
             }
         }
     }
